requirements.txt
pandas
numpy
scikit-learn
torch
plotly_express